[Bug] open notification in Different chains

+++ app/src/main/java/io/goldstone/blockchain/module/common/tokendetail/tokendetail/presenter/TokenDetailPresenter.kt
@@ -219,7 +219,12 @@ class TokenDetailPresenter(
 			val chartArray = arrayListOf<ChartPoint>()
 			val charCount = if (size > maxChartCount) maxChartCount else size
 			forEach {
-				chartArray.add(ChartPoint(CryptoUtils.dateInDay(it.date), it.balance.toFloat()))
+				chartArray.add(
+					ChartPoint(
+						CryptoUtils.dateInDay(it.date),
+						it.balance.toBigDecimal().toFloat()
+					)
+				)
 				if (chartArray.size == charCount) {
 					header?.setCharData(chartArray.reversed().toArrayList())
 				}
